Finish stub; small cleanup and refactors
Squashed commit of the following:

commit 88dfbdc606657ac6441eeba52e0c8563b0da205c
Author: Rory Tyler Hayford <rory.hayford@protonmail.com>
Date:   Wed Apr 29 15:44:17 2026 +0700

    Formatting

commit 3f81ce33d285d2145d2a002726c0c2ee51ff9902
Author: Rory Tyler Hayford <rory.hayford@protonmail.com>
Date:   Wed Apr 29 15:39:38 2026 +0700

    Some cleanup

commit 1a6354bf47a2c20ba229696d1aea16eb0bfcf21a
Author: Rory Tyler Hayford <rory.hayford@protonmail.com>
Date:   Wed Apr 29 14:01:24 2026 +0700

    Move UF primitives to new module

commit 07b13c6c676b82bb98745891c0d6c00e14a84138
Author: Rory Tyler Hayford <rory.hayford@protonmail.com>
Date:   Wed Apr 29 13:39:11 2026 +0700

    Add `inferPossibleTypes`

commit 6ff48c0c69421fb99b01d8d2555cea29064d2769
Author: Rory Tyler Hayford <rory.hayford@protonmail.com>
Date:   Tue Apr 28 13:15:25 2026 +0700

    `inferTypeReps`, some small refactors

commit db819acc237d3894f8b62d06951e8ad2ed1d10cc
Author: Rory Tyler Hayford <rory.hayford@protonmail.com>
Date:   Mon Apr 27 15:36:10 2026 +0700

    Add `closeOverTypeReps`

commit 72e0ece4192965cf39836aef18024f77959c5ee5
Author: Rory Tyler Hayford <rory.hayford@protonmail.com>
Date:   Mon Apr 27 13:17:38 2026 +0700

    Just remove `Subst` and `apply`

commit 6c92df0385290a6793001a5b40e2aaf5c31a0be1
Author: Rory Tyler Hayford <rory.hayford@protonmail.com>
Date:   Mon Apr 27 12:50:06 2026 +0700

    Do NOT free UF store to `Subst`; essentially drop `Subst`

commit cf1b68bab2c70e1ccea46b3acfc414c20e1279dd
Author: Rory Tyler Hayford <rory.hayford@protonmail.com>
Date:   Fri Apr 24 18:33:57 2026 +0700

    Fix some efficiency issues

commit 6e7f86e51d2c60f37352a2e57c827d9b83f66361
Author: Rory Tyler Hayford <rory.hayford@protonmail.com>
Date:   Fri Apr 24 18:26:11 2026 +0700

    First attempt at `inferExpr`
