+++ inferno-types/src/Inferno/Types/Type.hs
@@ -41,7 +41,6 @@ import Inferno.Types.Syntax
     RestOfRecord (..),
     Scheme (..),
     Substitutable (ftv),
-    substMap,
     TCScheme (..),
     TV (..),
     TypeClass (..),
@@ -50,6 +49,7 @@ import Inferno.Types.Syntax
     namespaceToIdent,
     punctuate',
     sch,
+    substMap,
     tySig,
     typeBool,
     typeDouble,
